news_main/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.4
django-bootstrap4==22.3
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.2.2
django-js-asset==2.0.0
django-localflavor==3.1
django-markdownx==3.0.1
django-precise-bbcode==1.2.15
django-ranged-response==0.2.0
django-redis==5.2.0
django-rest-elasticsearch==0.4.2
django-simple-captcha==0.5.17
djangorestframework==3.14.0
easy-thumbnails==2.8.5
elasticsearch==7.9.0
elasticsearch-dsl==7.4.0
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-memcached==1.59
python-stdnum==1.18
python3-openid==3.2.0
pytz==2022.7
redis==4.4.2
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
